change gps type

+++ main.go
@@ -28,10 +28,10 @@ type device struct {
 	SiteName string `json:"SiteName"`
 	// Gps_fix float32 `json:"gps_fix"`
 	// Ver_app string `json:"ver_app"`
-	Gps_lat float32 `json:"gps_lat"`
+	Gps_lat float64 `json:"gps_lat"`
 	S_t0 float32 `json:"s_t0"`
 	Timestamp string `json:"timestamp"`
-	Gps_lon float32 `json:"gps_lon"`
+	Gps_lon float64 `json:"gps_lon"`
 	// Date string `json:"date"`
 	// Tick float32 `json:"tick"`
 	Device_id string `json:"device_id"`
